@@ -1,59 +1,51 @@
 {
-  "name": "React-starter-app",
+  "name": "react-frontend-starter",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "babel-eslint": "^8.2.1",
-    "cookies-js": "^1.2.3",
-    "core-decorators": "^0.20.0",
-    "core-js": "^2.5.3",
-    "debounce": "^1.1.0",
-    "eslint": "^4.16.0",
-    "eslint-plugin-jest": "^21.7.0",
-    "eslint-plugin-promise": "^3.6.0",
-    "eslint-plugin-react": "^7.6.0",
-    "fuzzy": "^0.1.3",
-    "image-extensions": "^1.1.0",
-    "lodash.throttle": "^4.1.1",
-    "moment": "^2.20.1",
-    "node-sass-chokidar": "0.0.3",
-    "npm-run-all": "^4.1.2",
-    "path": "^0.12.7",
-    "promise-polyfill": "^7.1.0",
-    "prop-types": "^15.6.0",
-    "react": "^16.2.0",
-    "react-addons-css-transition-group": "^15.6.2",
-    "react-dom": "^16.2.0",
-    "react-redux": "^5.0.6",
-    "react-responsive": "^4.0.4",
-    "react-router-dom": "^4.2.2",
-    "react-scripts": "1.0.17",
-    "react-share": "^2.0.0",
-    "recompose": "^0.26.0",
-    "redux": "^3.7.2",
-    "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.2.0",
-    "request": "^2.83.0",
-    "stylelint": "^9.1.1",
-    "stylelint-config-standard": "^18.2.0",
-    "stylelint-processor-html": "^1.0.0",
-    "stylelint-scss": "^2.5.0",
-    "url-polyfill": "^1.0.11",
-    "uuid": "^3.2.1",
-    "whatwg-fetch": "^2.0.3"
+    "@craco/craco": "^5.2.4",
+    "@rebass/grid": "^6.0.0",
+    "@storybook/react": "^5.1.9",
+    "antd": "^3.20.7",
+    "apollo-boost": "^0.1.9",
+    "apollo-cache-inmemory": "^1.5.1",
+    "apollo-client": "^2.5.1",
+    "apollo-link": "^1.2.8",
+    "apollo-link-http": "^1.5.11",
+    "craco-antd": "^1.11.0",
+    "craco-less": "^1.9.0",
+    "graphql": "^14.1.1",
+    "graphql-tag": "^2.10.1",
+    "react": "^16.4.1",
+    "react-apollo": "^2.1.5",
+    "react-dom": "^16.4.1",
+    "react-router-dom": "^4.3.1",
+    "react-scripts": "^3.0.1",
+    "styled-components": "^4.1.2",
+    "webfontloader": "^1.6.28"
+  },
+  "devDependencies": {
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-config-prettier": "^2.9.0",
+    "eslint-plugin-import": "^2.11.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-prettier": "^2.6.0",
+    "eslint-plugin-react": "^7.7.0",
+    "prettier": "^1.13.5"
   },
   "scripts": {
-    "build-css": "node-sass-chokidar src/ -o src/",
-    "watch-css": "npm run build-css && node-sass-chokidar src/ -o src/ --watch --recursive",
-    "start-js": "react-scripts start",
-    "start": "npm-run-all -p watch-css start-js",
-    "build-js": "react-scripts build",
-    "build": "npm-run-all build-css build-js",
-    "test": "react-scripts test --env=jsdom",
+    "start": "craco start",
+    "build": "craco build",
+    "test": "craco test",
+    "lint": "eslint \"src/**/*.js\"",
     "eject": "react-scripts eject",
-    "js-lint": "eslint \"src/**/*.js\" \"public/**/*.js\"",
-    "css-lint": "stylelint \"src/**/*.scss\" \"public/**/*.scss\"",
-    "lint": "npm-run-all js-lint css-lint"
+    "storybook": "start-storybook -p 9009 -s public",
+    "clean": "rm -rf node_modules && npm i"
   },
-  "devDependencies": {}
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
