@@ -1,4 +1,8 @@
-module.exports = {
-  apiUrl: process.env.REACT_APP_API_URL || 'api-url-here',
-  siteURL: process.env.REACT_APP_URL || 'site-url-here',
+export default {
+  graphqlUrl: process.env.REACT_APP_API_URL || 'http://localhost:5000/graphql',
+  socialMediaLinks: {
+    facebook: 'https://www.facebook.com',
+    instagram: 'https://www.instagram.com',
+    linkedin: 'https://www.linkedin.com'
+  }
 }
