@@ -1,7 +1,3 @@
-import Notification from './Notification'
-import NotificationList from './NotificationList'
-
-export {
-  Notification,
-  NotificationList,
-}
+export { default as Footer } from './Footer'
+export { default as NavBarSignedIn } from './NavBarSignedIn'
+export { default as NavBarSignedOut } from './NavBarSignedOut'
