@@ -1,17 +1,5 @@
-import './lib/polyfills'
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { BrowserRouter } from 'react-router-dom'
-import App from './app/App'
-import registerServiceWorker from './registerServiceWorker'
-import { Provider } from 'react-redux'
-import store from './redux/store'
+import App from './App'
 
-ReactDOM.render(
-  <Provider store={store}>
-    <BrowserRouter>
-      <App/>
-    </BrowserRouter>
-  </Provider>
-  , document.getElementById('root'))
-registerServiceWorker()
+ReactDOM.render(<App />, document.getElementById('root'))
